server: remove module indirection
